Added Xbox node controlland fix node panel

+++ can_devices/encoder/encoder/encoder_rm.py
@@ -3,7 +3,7 @@ import can
 import time
 import rclpy
 from rclpy.node import Node
-from std_msgs.msg import Int32, Int16
+from std_msgs.msg import Int32, Int16, Float64
 
 class EncoderPublisher(Node):
 
@@ -11,7 +11,7 @@ class EncoderPublisher(Node):
         super().__init__('encoder_rm')
         self.publisher_absolute_pos = self.create_publisher(Int32, 'encoder/abs_pos', 10)
         self.publisher_turn_number = self.create_publisher(Int16, 'encoder/turn_num', 10)
-        self.publisher_angle = self.create_publisher(Int16, 'encoder/angle', 10)
+        self.publisher_angle = self.create_publisher(Float64, 'encoder/angle', 10)
         timer_period = 0.01 #Seconds
         self.timer = self.create_timer(timer_period, self.timer_callback)
         self.bus = can.interface.Bus(bustype='socketcan', channel='can1', bitrate=500000)
@@ -22,7 +22,7 @@ class EncoderPublisher(Node):
     def timer_callback(self):
         absPos = Int32()
         turNum = Int16()
-        angle = Int16()
+        angle = Float64()
         self.bus.send(self.sendMsg, timeout=1)
         receivedMsg = self.bus.recv(1)
         if receivedMsg is not None:
@@ -33,7 +33,7 @@ class EncoderPublisher(Node):
             absPos.data = decimalPos
             mod = decimalPos%self.steps
             turNum.data = decimalPos//self.steps
-            angle.data = int((mod*self.degrees)/self.steps)
+            angle.data = float((mod*self.degrees)/self.steps)
             self.publisher_absolute_pos.publish(absPos)
             self.publisher_turn_number.publish(turNum)
             self.publisher_angle.publish(angle)
